requirements.txt
tensorflow
gradio
pillow
numpy
